requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
ultralytics>=8.0.0
RPi.GPIO>=0.7.0
paho-mqtt==1.6.1 